Additions, fixes

+++ FFXIVCraftingSim/Types/GameData/LevelDifferenceInfo.cs
@@ -8,8 +8,8 @@ namespace FFXIVCraftingSim.Types.GameData
 {
     public class LevelDifferenceInfo
     {
-        public short Difference { get; set; }
-        public short ProgressFactor { get; set; }
-        public short QualityFactor { get; set; }
+        public int Difference { get; set; }
+        public int ProgressFactor { get; set; }
+        public int QualityFactor { get; set; }
     }
 }
